+++ app/entity/model/Genre.go
@@ -1,7 +1,7 @@
 package model
 
 type Genre struct {
-	GenreId uint   `gorm:"primaryKey" json:"genre_id"`
+	GenreId int64  `gorm:"primaryKey" json:"genre_id"`
 	Name    string `gorm:"not null" json:"name"`
 	Slug    string `gorm:"index:,not null" json:"slug"`
 	Status  int    `gorm:"index:,default:1" json:"status"`
