+++ app/entity/model/MovieCountry.go
@@ -1,6 +1,6 @@
 package model
 
 type MovieCountry struct {
-	MovieId   uint `gorm:"index:,not null" json:"movie_id"`
-	CountryId uint `gorm:"index:,not null" json:"country_id"`
+	MovieId   int64 `gorm:"index:,not null" json:"movie_id"`
+	CountryId int64 `gorm:"index:,not null" json:"country_id"`
 }
