+++ app/entity/model/MoviePeople.go
@@ -1,6 +1,6 @@
 package model
 
 type MoviePeople struct {
-	MovieId  uint `gorm:"index:,not null" json:"movie_id"`
-	PeopleId uint `gorm:"index:,not null" json:"people_id"`
+	MovieId  int64 `gorm:"index:,not null" json:"movie_id"`
+	PeopleId int64 `gorm:"index:,not null" json:"people_id"`
 }
