+++ app/entity/request/EpisodeDetailRequest.go
@@ -3,6 +3,6 @@ package request
 type EpisodeDetailRequest struct {
 	Name          string `json:"name"`
 	Link          string `json:"link"`
-	EpisodeTypeId uint   `json:"episode_type_id"`
+	EpisodeTypeId int64  `json:"episode_type_id"`
 	Status        int    `json:"status"`
 }
