+++ pkg/validator/MovieCountryValidator.go
@@ -4,7 +4,7 @@ import (
 	"github.com/xdorro/golang-fiber-base-project/app/entity/model"
 )
 
-func ExistCountryIdInCountries(countryId uint, countries []model.Country) bool {
+func ExistCountryIdInCountries(countryId int64, countries []model.Country) bool {
 	for _, country := range countries {
 		if country.CountryId == countryId {
 			return true
@@ -14,7 +14,7 @@ func ExistCountryIdInCountries(countryId uint, countries []model.Country) bool {
 	return false
 }
 
-func ExistCountryIdInCountryIds(countryId uint, countryIds []uint) bool {
+func ExistCountryIdInCountryIds(countryId int64, countryIds []int64) bool {
 	for _, country := range countryIds {
 		if country == countryId {
 			return true
